feat: add camera icon
