+++ router/router.jsx
@@ -4,10 +4,10 @@ import { HomePage } from "../src/pages/Home.page";
 import { ProductDetailPage } from "../src/pages/ProductDetail.page";
 import { RentPage } from "../src/pages/Rent.page";
 import { SigninPage } from "../src/pages/Signin.page";
-import { SignupPage } from "../src/pages/Signup.page";
-import { AccountPage } from "../src/pages/Account.page";
-import { SchoolPage } from "../src/pages/School.page";
-import { SchoolCardPage } from "../src/pages/SchoolCard.page";
+import { SignupPage } from "../src/pages/signup/Signup.page";
+import { AccountPage } from "../src/pages/signup/Account.page";
+import { SchoolPage } from "../src/pages/signup/School.page";
+import { SchoolCardPage } from "../src/pages/signup/SchoolCard.page";
 import { PayMoneyPage } from "../src/pages/PayMoney.page";
 import { MyPage } from "../src/pages/My.page";
 import { ChargePage } from "../src/pages/Charge.page";
