+++ src/components/common/Icon.component.jsx
@@ -5,6 +5,7 @@ import { RiHome5Line } from "react-icons/ri";
 import { AiOutlineClose } from "react-icons/ai";
 import { VscAccount } from "react-icons/vsc";
 import { CiViewList } from "react-icons/ci";
+import { AiOutlineCamera } from "react-icons/ai";
 import classnames from "classnames";
 
 const IconSize = {
@@ -32,6 +33,7 @@ export const Icon = ({ type, size = "medium", className }) => {
     account: <VscAccount className={combileClassName} />,
     viewlist: <CiViewList className={combileClassName} />,
     message: <BiMessageRoundedDetail className={combileClassName} />,
+    camera: <AiOutlineCamera className={combileClassName} />,
   };
 
   return IconType[type];
