+++ src/components/sign/Information.component.jsx
@@ -8,13 +8,13 @@ import { Txt } from "../common/Txt.component";
 export const SignInformation = ({ title }) => {
   return (
     <>
-      <Txt typography="h6" colors="black" className="my-20">
+      <Txt typography="h6" colors="black" className="my-10">
         안녕하세요!
       </Txt>
       <Txt
         typography="h3"
         colors="black"
-        className="flex flex-col gap-2 mb-20 text-center"
+        className="flex flex-col gap-2 mt-5 mb-20 text-center"
       >
         {title.split("\n").map((line) => {
           return <div>{line}</div>;
